+++ extensions/minigraph-playground-engine/src/test/java/com/accenture/minigraph/playground/PlaygroundTest.java
@@ -111,7 +111,6 @@ class PlaygroundTest {
     static void setup() {
         deleteTempGraph();
         AutoStart.main(new String[0]);
-
     }
 
     @AfterAll