+++ system/event-script-engine/src/main/java/com/accenture/services/plugins/logical/IsNullOperator.java
@@ -31,10 +31,7 @@ public class IsNullOperator implements PluginFunction {
 
     @Override
     public Object calculate(Object... input) {
-        if (input == null) {
-            return true;
-        }
-        else if (input.length == 1) {
+        if (input.length == 1) {
             return input[0] == null;
         }
         throw new IllegalArgumentException("Only one value is accepted");